Prevent to load fonts from ass_set_fonts_dir with Direct Write provider
"ass_set_fonts_dir" currently not support to handle hundreds of fonts as Kodi would like to do,
to load fonts from the Kodi "fonts" user folder.

When the fonts path is set with "ass_set_fonts_dir" method
these fonts will be loaded by default with a "standard" method
that put in RAM all fonts in the folder,
this will fill the system RAM causing a system crash.

Here we have disabled the "standard" method on Windows case only
because currently is the only provider that allow to load from a custom
folder, and is the only OS where we provide a patched library.

What happens on Windows with Direct Write provider:
-DW load the fonts from the system
-DW support to load the fonts from a custom folder
Results that an user can use the Kodi user "fonts" folder without side effects, but the loading could be slow in any case.
If there are too much fonts like 500,1000 or higher, suggest to put these fonts in the system this could improve the loading (untested).

On all other OS we do not have the patched libass,
then the loading of fonts in the custom folder (user folder) use the "standard" bugged method,
the only workarounds are explained below.

Linux with FontConfig provider:
-FC load the fonts from the system
-Load the fonts from user folder can be done only via "standard" method
Suggest the users to install fonts in the system,
or few fonts in the user folder.

Kodi OS like LibreElec/CoreElec/... with FontConfig provider:
-FC load the fonts from the system
-Load the fonts from user folder can be done via "standard" method (or see alternatives)
Suggest the users to install fonts in the system, or few fonts in the user folder or see alternatives.
Alternative 1:
Allow the users to put fonts in another directory.
FontConfig need to be configured to load fonts from another custom
folder, this folder do not have to match to the Kodi "fonts" user
folder, or the fonts will be loaded twice!
Alternative 2:
Add a patch to Kodi sources, in order to comment the execution of
ass_set_fonts_dir method. In this way is possible use the alternative 1
by allowing to use the Kodi "fonts" user folder without side effects.

MacOS with CoreText provider:
-FC load the fonts from the system
-Load the fonts from user folder can be done only via "standard" method
Suggest the users to install fonts in the system,
or few fonts in the user folder.

Android is without font provider:
-Load the fonts from user folder can be done only via "standard" method
Suggest the users to add few fonts in the user folder.